+++ src/embedGoogleMapEmbedAPIHtmlBuilder.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * @version	$Id: Embed Google Map v2.2.0 2019-08-17 09:10 $
+ * @version	$Id: Embed Google Map v2.3.0 2019-12-30 11:19 $
  * @package	Joomla 1.6
  * @copyright	Copyright (C) 2014-2019 Petteri Kivimäki. All rights reserved.
  * @author	Petteri Kivimäki