+++ src/embedGoogleMapHtmlBuilder.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * @version	$Id: Embed Google Map v2.2.0 2019-08-17 09:10 $
+ * @version	$Id: Embed Google Map v2.3.0 2019-12-30 11:19 $
  * @package	Joomla 1.6
  * @copyright	Copyright (C) 2014-2019 Petteri Kivimäki. All rights reserved.
  * @author	Petteri Kivimäki
@@ -50,11 +50,12 @@ abstract class EmbedGoogleMapHtmlBuilder {
     }
 
     protected function getIFrameBegin(&$params) {
+        $title = "title='" . $params->getTitle() . "'";
         $width = "width='" . $params->getWidth() . "'";
         $height = "height='" . $params->getHeight() . "'";
         $style = "style='border: " . $params->getBorder() . "px " . $params->getBorderStyle() . " " . $params->getBorderColor() . "'";
         $styleClass = "class='embedGoogleMap'";
-        return "\n<iframe $styleClass $width $height $style ";
+        return "\n<iframe $title $styleClass $width $height $style ";
     }
 
     protected function getLinkHtml($url, $label) {