+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.3.0 - 2019-12-30
+- Add title parameter.
+  - Can be used to define the `title` attribute of the map's `iframe`. 
+
 ## 2.2.0 - 2019-08-17
 - Add link position parameter.